Create a 2 player  Simple Shooting Game using Html and Javascript
Create a 2 player shooting game. Each player has a health of 100. Players can shoot at each other with a random power (0-5) which will reduce the opponent's health. The Player whose health reaches 0, dies and the game ends. There are 5 such rounds, the player with the most game wins, wins the match/tournament. On the screen, there should be a button to start the game. Once the game finishes, it should update the current score. When a player wins 3 games, show the winner of the match.